5 : Split up the test files for parallel compilation
